:building_construction: checkpoint fallback initial impl
