+++ config/src/networks.rs
@@ -1,7 +1,7 @@
-use serde::Serialize;
-
-use crate::{bytes_serialize, ChainConfig, Fork, Forks};
 use common::utils::hex_str_to_bytes;
+
+use crate::base::BaseConfig;
+use crate::types::{ChainConfig, Fork, Forks};
 
 pub enum Network {
     MAINNET,
@@ -15,20 +15,6 @@ impl Network {
             Self::GOERLI => goerli(),
         }
     }
-}
-
-#[derive(Serialize, Default)]
-pub struct BaseConfig {
-    pub rpc_port: u16,
-    pub consensus_rpc: Option<String>,
-    #[serde(
-        deserialize_with = "bytes_deserialize",
-        serialize_with = "bytes_serialize"
-    )]
-    pub checkpoint: Vec<u8>,
-    pub chain: ChainConfig,
-    pub forks: Forks,
-    pub max_checkpoint_age: u64,
 }
 
 pub fn mainnet() -> BaseConfig {
